Added some feature detail page and refactor code
