+++ PokeDex/Constant/AppFonts.swift
@@ -10,8 +10,10 @@ import UIKit
 
 struct AppFonts {
     static let placeholderFont = UIFont(name: "Poppins-Regular", size: 15)
-    static let titleBoldFont = UIFont(name: "Poppins-Bold", size: 30)
+    static let titleBoldFont = UIFont(name: "Poppins-Bold", size: 28)
     static let regulerBoldFont = UIFont(name: "Poppins-Bold", size: 16)
     static let cardIdFont = UIFont(name: "Poppins-Regular", size: 12)
     static let cardNameFont = UIFont(name: "Poppins-Regular", size: 15)
+    static let smallFont = UIFont(name: "Poppins-Light", size: 12)
+    static let mediumFont = UIFont.systemFont(ofSize: 12, weight: .medium)
 }
