+++ PokeDex/Info.plist
@@ -23,6 +23,7 @@
 	<array>
 		<string>Poppins-Bold.ttf</string>
 		<string>Poppins-Regular.ttf</string>
+		<string>Poppins-Light.ttf</string>
 	</array>
 </dict>
 </plist>
